+++ dialogs/about.py
@@ -34,7 +34,7 @@ class AboutDialog(QDialog):
         layout.addWidget(title)
 
         # Version
-        version = QLabel(self.tr("Version 1.0.0-rc.7"))
+        version = QLabel(self.tr("Version 1.0.0-rc.8"))
         version.setAlignment(Qt.AlignCenter)
         layout.addWidget(version)